requirements.txt
flet[desktop]
pyinstaller==6.12.0
pdfplumber==0.11.5
PyPDF2==3.0.1
pdfminer.six==20231228